ssfn2: moved to brutal/codec.
