+++ sources/libs/brutal/codec/ssfn2/ssfn2.h
