+++ sources/pkgs/fontinfo/main.c
@@ -1,6 +1,6 @@
 #include <brutal/alloc.h>
+#include <brutal/codec/ssfn2/ssfn2.h>
 #include <brutal/debug.h>
-#include <brutal/gfx/ssfn2/ssfn2.h>
 #include <brutal/io.h>
 
 int main(int argc, char const *argv[])
